Synchronize WindowInputTests#testMoveWindowAndTap
After moving the position of a window, there is no guarantee that
InputDispatcher has received the bounds update unless we synchronize
input transactions, after which we will know for sure that the bounds
update has propagated to input.

We don't use the WindowInfos callback because that gives the bound in
unrotated coordindates, and we don't want to add complexity for
calculating the rotated bounds to ensure that the bounds have been
updated.

Bug: 324038855
Bug: 295885275
Test: atest WindowInputTests
Change-Id: Ia72ed7afac959f841dfd1dc4256f9d03457a47bf