correct sdk_version of TestApp_JavaProtos
This change corrects sdk_version for TestApp_JavaProtos. Before the
sdk_version is not set, and by default the library will be built against
the implementation of the framework. TestApp_JavaProtos does not need to
build against the implementation of the framework. Set the sdk_version
to the right scope.

Test: presubmit
Bug: 320485496
Change-Id: I4e8292e06a1c5a900cf442bccf61a893d1997abd