Attempt to stabilize InputMethodServiceSwitchingTest
This is a follow up CL to my previous CL [1], which introduced
InputMethodServiceSwitchingTest as part of migration from host-side
tests to device-side tests.

One thing I overlooked is that there is a delay between installing an
IME APK and InputMethodManagerService (IMMS) recognizes it.  With this
CL the test explicitly waits until the test IME is recognized by IMMS
to make the test deterministic.

 [1]: I904854a2d6bb7324748cc20ba2de3b0a2b0db4b7
      dc71ce46c95ba1e3469c07187b89133d0786ebbb

Bug: 323251870
Fix: 328267676
Test: atest CtsInputMethodServiceHostTestCases
Change-Id: Ia2391af1d0f98ac291f7b4c2a166cda66ed08d8d