Take screenshot upon failure in WindowInputTests
The last attempt to take screenshots upon failure did not succeed. This
is a new attempt. This was tested locally by adding an intentional
failure and checking for the artifacts in the local temp dir:

file:///tmp/atest_result_svv/20240306_102157_yg8_u8lc/log/test_logs.html

Confirmed that the png was present in the local test run.

Bug: 324038855
Test: atest WindowInputTests
Change-Id: I99662e3337c48e43f222b92a468aa98fde32cb05