Ignore test if enterprise profile cannot be set up
This is a guess as to the cause of the related bug which is that the
enterprise profile cannot be found or set up for some reason.

Ideally, we don't just unilaterally skip the test if we can't set up
the enterprise profile but find out why the profile cannot be set up
in some cases hence the added logging.

Bug: 322917418
Test: EnterpriseContactsMultiUserTest.java
Change-Id: I91af8702a4d953b9e15cec6f6a75356b64672596