CtsWindowManagerDeviceTestCases: Test fails when Toast is moved to top of display
When the toast moved to top of display, the touch is not received by the
activity content area as the action bar is present behind the toast but not the activity content.

Removing the action bar from the activity, so that the activity content area will move up and the test passes.

Bug: 312839622
Test: Executed the CtsWindowManagerDeviceTestCases module test cases
Change-Id: I8b5e4e3d80b2ec296d737a3cec878828bd621231