Release hanging pointers in StylusHandwritingTests
This test suite uses injectStylusDownEvent, and injectStylusUpEvent to
inject the required MotionEvents. However in some cases, we get the
ACTION_DOWN event, but never get to inject the ACTION_UP event, which
leaves a hanging pointer. The test suite also uses global injection, so
the hanging pointer will create test failrues for all the test runs that
follow it, and that use targeted input injection.

This adds the missing injectStylusUpEvent calls, and also wraps the code
between the DownEvent and UpEvent in a try-finally block, so failed
assertions still have the chance to clean up the hanging pointer.

Bug: 325928186
Test: atest StylusHandwritingTest
Change-Id: I43c209e1ecaf9c2fb05bfb13c1f2fa048538c64e