ITS: add physical cam to manual testing session
Details: when running ITS manually (without a tablet), a session is
created to confirm that the scene is OK. However, hidden physical camera
IDs are not handled, and '0.x' is passed as the logical camera. Changed
behavior to handle these strings and parse the hidden physical ID, if
it's available.

Bug: 328609269
Test: tested locally
Change-Id: Ifb6d65ff6602aafad724b5070c38d915647246b3