Mark testImeRemainsVisibleAfterSwitchingIme flaky
This is a follow up CL to my previous CL [1], migrated the regression
test for Bug 152876819 from

  CtsInputMethodServiceHostTestCases

to

  CtsInputMethodTestCases

with hoping that the known test flakiness (Bug 324493679) might be
improved if the test runs as a device-side test.

Unfortunatelly the test remained unstable as seen in Bug 328694633,
and this could be a real system issue rather than a test issue.

Anyway let's mark the test with @FlakyTest annotation so as not to
block other CLs.

 [1]: Idf2a3b0f70277815cac7c78d0bb0adf9a505ab77
      3ebd32c02c6853f1c0cda52b3d15a6cd0cdad4c0

Bug: 328694633
Test: presubmit
Change-Id: Id00917072e90f3eb17276bf1f0a5456ce725fdad