Disable SecureFrpInstallTest in CTS given FRP enforcement
The existing test relies on the fact that the user can override the FRP
state from shell.  This is one of the way an attacker can use to bypass
FRP, and is now prevented by the FRP hardening work.

Disable the test when the feature flag is on. The ground truth of the
FRP state can't be change from shell, not on user build.

Bug: 303188712
Test: atest SecureFrpInstallTest (skipped when the flag is on)
Change-Id: I75e844955c807198015bae06f371ece9ce26395a