Change order of Preparers to set feature flags first and install apk
The order is changed to set the feature flag first and then install
the test APKs. If the test device's feature flags are different from
the feature flags defined in FeatureFlagTargetPreparer, a race
condition can occur when rebooting without completing test APK
installation and permissions. Therefore, change the order to install
the APK after setting the feature flag.

Bug: 322750991
Test: atest CtsTelephonyTestCases
Change-Id: Ia569b7c48e0e657cc26bfa4e74e101fabd2e35a3