Do not limit nav bar numbers when check existance
There are some tests trying to verify the existance of navigation bar on
the display. We were using the window type to identify the navigation
bar and it was only one possible navigation bar one each display.
However, given the flexible insets change, we generalized the test to
use the insets provided by a window to identify the navigation bar
windows, and we don't have limit of how many windows can provide the
navigation bars.

Lift the limitation when check the existance, in the meanwhile,
refactored a little bit to make the methods clean.

Test: MultiDisplaySystemDecorationTests#testNavBarShowingOnDisplayWithDecor
Bug: 328374159
Change-Id: Ic60b05abd5fdf2d21989d82b0f9561ba7edbc89b