E2E CTS Integration test of simultaneous calling restrictions.
This uses CTS MockModem infrastructure and IMS configurations
to set the enabled logical slots that is being returned from the modem.
This test then uses ImsServiceConnector to ensure register IMS
for each of the 2 subscriptions. Using these variables, we
ensure that the simultaneous calling restrictions are being set properly for each phone account.

Bug: 328293858
Test: SimultaneousCallingRestrictionsTest
Change-Id: I484000515e0373043866e568b42044c034a0ca91