Choose preferred application
The changes are intended to handle two flows when the chooser dialog displays preferences

Case 1: The chooser dialog displays a list of activity names. Provide the activity name to get the UiObject from the dialog

Case 2: The chooser dialog displays a list of applications. By providing the application's label, UiObjects can be retrieved. After that, pass the activity name to retrieve the UiObject

Bug: 307365381
Test: atest android.dynamicmime.cts.PreferredActivitiesTestCases
Change-Id: I5e6c4fa540b871ccd7c7ce93f6373c10c8ea6da3
Signed-off-by: Manjunath <m.rathod@samsung.com>