Enable VP9 profile tests
Enable the value pair of profile and level information for VP9 clips.
Some VP9 clips may not contain the level information in CSD or may not
have a CSD at all. In such cases, profile can be synthesized from the
first frame, but level cannot be. So, this change adds a check to verify
the level only if the CSD contains it.

This relands f1bebdc7a23abb60202235d5c8d88a10032c9892

Test: atest CtsMediaV2TestCases:ExtractorTest
Bug: 159582475
Bug: 295804596
Change-Id: I5c5bfe4acab3a2a52c0c21cb4ed13dfce2719b81