mediav2: Limit block model test to Android V and above
CodecDecoderTest was updated to test block model mode after Android U
was released. This part is not expected to work as expected on older
versions, as some of the fixes needed for this were merged in the
framework after Android U release. Hence limit this part of the tests
to devices running Android V and above. This is needed to run MTS tests
on older versions of Android.

Bug: 329767811
Bug: 328764199
Test: Ensure the following doesn't show any unexpected failures when
testing devices running Android U.
 atest CtsMediaV2TestCases --  --module-arg CtsMediaV2TestCases:\
instrumentation-arg:codec-prefix:=c2.android

Change-Id: I4725c1fbd3abaf521ad0235d90ab756cb28f087a