Test usage of PermissionManager in BroadcastSkipPolicy
We are fixing a bug in BroadcastSkipPolicy, where default-appop mode for
an appop permission was being considered as invalid for broadcast
delivery. We're here adding a new test that would verify the fix (the
test passes without the fix, and passes with the fix).

In addition to the new test targetting the default appop mode, we're
also adding tests for the allowed and errored appop modes.

Bug: 315468967
Test: atest ContextTest

Change-Id: I08ecf2ff95bce1851da38f42a61e4161ba313543