Fix flake in activeScanRouteDiscoveryPreference_scansOnSelfScanProvider
The change moves the test to the hostside test suite to avoid other
tests from binding first to the service, and causing the actual test to
not receive the event. Hostside tests normally use one process per test
and then discard it.

See b/327120745#comment10 for more details about the root cause of the
test flakiness.

Bug: 327120745
Test: atest MediaRouter2HostSideTest
Change-Id: I1aa3a6b10f00334265db245e7616bc17e5469149