Fix flaky preferred mixer attributes test
When calling the mixer attributes API (set and clear) the callback
sometimes happens right away and test callback was not recieved at the
test. Changed the test so that a callback is registered before calling
the APIs, this way it is assure that the callback will be triggered
when the audio service triggers it from the API calls.
Also changed the timeout units, as the original constant is set to 5
seconds, as suggested by the name. A 5 milliseconds timeout was leading to a
failure rate of about 1/6 in automotive devices. No new flakes were
observerd in 100 runs.

Bug: 329561394
Test: atest testPreferredMixerAttributes --rerun-until-failure 100
Change-Id: I5c2267fa18f37678e84a95327450fb88b817f164