Replace main thread executor with ExecutorService in CarWatchdogManagerTest
Before this change the following a race condition existed primarily in testCheckHealthStatus.
The test would pass correctly but the executor would still be executing after the test case had exited.
In the tearDown method, permissions are removed. Since the executor was executing a code path that required permissions,
a SecurityException was thrown as the test had its permissions removed at that point.

Use an ExecutorService instead and wait for the executor to finish executing tasks before exiting the test.

Test: atest CarWatchdogManagerTest
Fixes: 326676871
Change-Id: If2b014e8f1ac56a60130318897b770304012ce4f