AppStartInfo CTS Tests: Add tests for 'current' process and timestamps.
This adds a test to query the ApplicationStartInfo from the 'current'
process. Currently, that doesn't work and the test validates that it's
unsuccessful. I'm not sure if that's expected and filed b/330961867 to
follow up.

This also adds a test that adding a timestamp to the
ApplicationStartInfo object and immediately retrieving it works
properly. I've filed b/330962538 to confirm whether those timestamps
should be available in a subsequent ApplicationStartInfo query.

Last, this disables testAppRemoved for now because uninstalling the
test app breaks other tests that were looking for it. Whoops.

Bug: 323464261
Bug: 325776365
Bug: 281761565
Test: atest CtsAppStartTestCases
Change-Id: I26b82f16f9f34a73eff7702cde48cf6930c59fb6