Extract getTestMarker and require EditText tag
This extracts the getTestMarker defined in many subclasses of
EndToEndImeTestBase to the superclass, and adds a required EdiTextTag
parameter, so we can tell apart the EditTexts in test cases where
multiple of them are present at the same time.

Also make ExtractViewTest a subclass of EndToEndImeTestBase to make use
of this method, which should provide the setup and teardown methods we
use in our other test suites as well.

Bug: 331194037
Test: atest
  AccessibilityInputMethodTest
  EditTextImeSupportTest
  ExtractViewTest
  FocusHandlingTest
  ImeInsetsVisibilityTest
  InProcessImeTest
  InputConnectionEndToEndTest
  InputConnectionHandlerTest
  InputConnectionLifecycleTest
  InputMethodStartInputLifecycleTest
  KeyboardVisibilityControlTest
  OnScreenPositionTest
  PackageVisibilityTest
  StylusHandwritingTest
  InputMethodManagerTest
  EndToEndImeTestBase
Change-Id: Idbd0c7c7b6b0ba7576767d962d352b4edf39e09e