PointerIconTest: Add tests for drawing tablet
Extend PointerIconTests to be parametric tests with the parameter being
an input device type. This way, we can run the tests for different
devices such as mice, touchpads, and drawing tablets.

This CL adds support for drawing tablets.

We also update the existing icon hotspots so that we are no longer using
fractional hotspots. When using a fractional hotspot, the compositor can
generate artifacts that results in different rendering on different
devices.

Bug: 331194858
Test: atest PointerIconTest
Change-Id: Ic09720b309d88e87ddf1c6f91875ec967bced65e