Disable Sony Dualshock 4 Touchpad tests on Android 14
The way the touchpad is processed for Sony Dualshock 4 controllers is
changing in a way that changes the event stream that is produces.
Disable the CTS test that enforces the old event stream in Android 14.

The test will be updated in a future release.

Bug: 329585708
Test: atest SonyDualshock4BluetoothTest
Change-Id: I293d5ccf7f8011a9d7a9a1300c083601711d0ef8
Merged-In: I93c6a9abba86334cfd3583aecaa4f7d8fb90372c