Add test coverage for key behavior over dreams.
This changelist introduces a new set of tests to exercise key
event behavior over dreams. These tests are implemented using
a new test dream, which can be directly modified during test
execution. This allows dream characteristics to be modified,
such as which keys to consume and the interactive nature of
the dream.

Bug: 326975875
Test: atest DreamServiceTest#testKeyHandling_InsecureKeyguardDismissesOnConfirmKey
Test: atest DreamServiceTest#testKeyHandling_SecureKeyguardConfirmKeyPromptsBouncer
Test: atest DreamServiceTest#testKeyHandling_InteractiveDreamConsumesConfirmNoWakeup
Test: atest DreamServiceTest#testKeyHandling_InteractiveDreamDoesNotConsumeConfirmPromptsBouncer

Change-Id: I7d8e6bd855a12c46d662a0922c80f7582671808b