mediav2 CTS: Enable output validation for encoder surface tests
The test queues compressed data to decoder. The decoded frames are
released to encoder for encoding. The output from encoder is muxed
for further analysis. The decoder-encoder pair are not free running.
The test stalls decoder from receiving more inputs until encoder has
sent output for earlier released frames. So there should be no
frame-drops.

Force the test to check if decoder input count and encoder output count
are same and their timestamps to be identical

Bug: 153127506
Test: atest android.mediav2.cts.CodecEncoderSurfaceTest

Change-Id: Ia18aa6a1c89df14b05be0e43a779c3be14dfbdc2