Skip zen tests on automotive platform
The tests are changing the streamVolume and expecting ringer mode
changes on automotive. These actions are not well defined when the
volume groups are containing multiple streams that do not have the same
aliasing system as the legacy stream volumes. This is the case for
seahawk devices for which the volume group telringvol contains ring,
voice and sco streams.

Test: trivial skip
Bug: 328734482
Change-Id: Ia94e38967eae6d865f0f1ecdb106aba57e619461