Correct sdk_version for CtsAttentionServiceDevice and CtsSoundTriggerInstrumentation library
This change set sdk_version as "test_current" for
CtsAttentionServiceDevice and CtsSoundTriggerInstrumentation. Before the
sdk_version is not set, and the libraries will be build against the
implementation of the framework jars.

The "test_current" sdk_version is selected, because CtsSoundTriggerTestCases uses TestApi in code,
and CtsAttentionServiceDevice depends on libaries have test_current.

Test: presubmit
Bug: 327965477
Change-Id: I17046bbd32a15c0c548aae1b697a1660fb4e9869