Add RequireRootInstrumentation.
Currently RequireAdbRoot only requires that adb has the capability of
gaining root - but this is not enough to ensure that we are able to use
many capabilities garded by root (which require the test to be
instrumented by a root shell). This annotation ensures that - and
recursively includes @RequireAdbRoot for inclusion in the rooted test
suites.

Also adds the -r/--root arguments to btest. We will unroot and exclude
@RequireAdbRoot from all tests when this is not included - and root and
include them when this is included.

Test: btest a.d.c.ApplicationHiddenTest -ss
Test: btest a.d.c.ApplicationHiddenTest -ss -r
Test: atest GtsRootDevicePolicyTestCases
Test: btest bedstead-root-tests -r
Fixes: 332757554
Change-Id: Id4e21b3bcb47a385118f2bf5f53250256ea87a0d