Remove 5 sec wait in tests that adds up to minutes.
I was running these tests and wondered why it took so long.
This profile showed the wait and the comments suggeste this could have
been fixed 6 years ago.

Looking at postsubmit profiles, I think this will shave off 5 minutes in
post submit.
I'm not sure if the user-switch wait is still needed or not, but I
shortened and extended the retries. Now may wait is 15 instead of 50
secs.

In my limited run (not all tests):
   before:  https://ui.perfetto.dev/#!/?s=1fc6e7aa4f43e7138f07968d8cd4e13dfde7bc296e5b1a180b56a0715e0eb54a
   after: https://ui.perfetto.dev/#!/?s=27602eddf7cb63ec6fc8f3f9c65908d62046487130254ac4f698470b2dfa4a1b

Bug:  b/114057686
Test: m CtsDevicePolicyManagerTestCases ; atest_tradefed.sh template/atest_device_test_base --template:map test=atest --module CtsDevicePolicyManagerTestCases  --no-early-device-release --module-arg CtsDevicePolicyManagerTestCases:include-annotation:com.android.cts.devicepolicy.annotations.PermissionsTest
Change-Id: I135238a9ce0136baa6ea7e21dd4d160a74ca8c50