Fixes CTS failure in ExtensionRearDisplayTest
Checks if the original state of the device would put the device
to sleep when the test transitions back to the original state.
If so, we should not check if the original activity was visible
again.

Fixes: 333314770
Test: ExtensionRearDisplayTest#testStartAndEndRearDisplaySession
Change-Id: I8a44ba06272f0b118dcbe9ad31b7b35d554ca867
Merged-In: I7d262710a1d88ef64b0dbc34073cb70dbff3d3d1
Merged-In: I8a44ba06272f0b118dcbe9ad31b7b35d554ca867