VCN CTS: Deflake tests by improving onLost verifications
The test tried to use #waitForLost to verify that a specific network
was reported onLost. However this verification logic is not always
reliable. All lost networks are stored in mLostHistory and #waitForLost
only returns the head of the mLostHistory. When mLostHistory has multiple
networks(due to onLost for the same network called multiple times or
multiple networks have matched the requests), the expected network
might not be in the head.

This patch fixes the issue by looking through all the networks in
mLostHistory instead of just the head.

Bug: 333924896
Test: atest VcnManagerTest for 100 times and all passed; without the
      fix the issue can be reproduced by running test for 20 times
Change-Id: Ie0a2da2bed5a8ebcfc5399b6211a324c318b4cca