Remove a not-so-useful test from InputMethodServiceTest
This CL removes

  testNoExceptionWhenSwitchingDisplaysWithImeReCreate,

which was added as a regression test against Bug 202081442 [1], as we
believe it's been causing more troubles than preventing issues in a
way we hoped.

There are a couple of issues in this test.

- The test is slow and flaky

Anecdotally implementing stress tests on top of MockIme ended up
creating flaky tests, and looks like this test is yet another example
of such stories. As shown in Bug 292824952, the test has costed us as
one of random flaky tests [2].

Note that we already have InputMethodStressTest [3].

- Stability reports from real devices have been more useful.

As shown in Bug 295880764, it is highly likely that we would receive
real reports from real IME processes running on real devices when

  android.inputmethodservice.SoftInputWindow#show()

throws an exception [4], while this test in question has never caught
any real issue since it was introduced as far as I am aware of.

With above, let's remove this flaky and not-so-useful test.

 [1]: I77364a4f5c1faf9a5189502a09fe1c4a8b303397
      1076a989eb0f9091c9aa72e3412f8b72bfff0f8b
 [2]: I94e3cdbbdc3483c3db95bd9fbd1352750fdbe6e9
      903a82ff5e3580560f8e2b69e3c1f8aa1a3fbe3c
 [3]: I1959225a7396d17750d799e11c42eb1899fe5bb9
      08603e695f9500893709b5ff746d1dff5b5d8320
 [4]: I66ba7b465e901e9ec2b31dd0245420513381872d
      8e6028eff3452c171ab6fae83fda05cc4e2a9630

Fix: 333620160
Test: atest CtsInputMethodTestCases:InputMethodServiceTest
Change-Id: Id095aa0c37d3a6ed779bb1765b2a857d5cbc8e58