DO NOT MERGE ANYWHERE Try checking against expected permissions for multiple releases
There are multiple releases between U and V which may have different
sets of permissions. Since the same test is used for all of them and
there's no proper way to distinguish between the variations we try
checking against a set of manifests where one of them will be compatible
with a correct set of permissions.

Test: Run this on all releases available
Bug: 320411042

Change-Id: Id2a503bd33eed0aecdc70c33832b8ad6dc65a60c