Fix Wear Compose View Search Behavior
Because UiScrollable#scrollForward() currently always returns false due
to a bug (b/306483780), the UiAutomatorUtils2#waitFindObjectOrNull API
never concluded that a scrollable is at the end (which would have
triggered a scroll back to the beginning of the scrollable to continue
search of the target view). This has caused some Views to never be found
if they are above the starting point of the search (because the
scrollable will never scroll up).

To mitigate this situation, this change adds a logic to always scroll up
a Wear Compose scrollable view before starting a search.

The following summarizes the old and new flows for Wear Compose:

Ideal behavior:
- search view
- exists?
  - No. is at end of scrollable?
    - Yes => scroll to top. Continue search
    - No => scroll down. Continue search
  - Yes. Done.

Behavior due to bug:
- search view
- exists?
  - No. is at end of scrollable?
    - No (always No because of bug) => scroll down. Continue search

Mitigation behavior (this CL):
- search view
- exists?
  - No. Has done one-time scroll to beginning?
    - Yes: scroll down => Continue search
    - No: scroll to top => Continue search
   - Yes. Done.

Bug: 330820852
Test: atest android.permission3.cts.PermissionTest23#testGrantPreviouslyRevokedWithPrejudiceShowsPrompt
Change-Id: I3ca0757d8dd9c86cf145c3acdf72abc4236496cd