Explicitly request focus on the views which should be initially focused
The original logic calls View#setFocusedByDefault to make the target
views be focused initially. However, it only takes effect when the
system is leaving touch mode. This CL explicitly requests focus on the
views which should be initially focused. So even if the focused views
are changed before running the test, the test will still properly
setup the initial focused views.

Fix: 334246829
Flag: NA
Test: atest ActivityEmbeddingFocusTests
Change-Id: Ie9a5641e0c1089a3062f23f206920fc5c96aad17