AdaptivePlaybackTest: Do not combine CSD with first frame for VP9
CSD in VP9 can't be combined with first frame as the data doesn't
have any start codes.
AdaptivePlaybackTest combines csd with first frame and queues it to
the encoder. This approach doesn't work for VP9 as elementary bitstream
in VP9 doesn't contain access delimiters like start codes.

This wasn't an issue so far as the extractor wasn't returning CSD, but
a recent change synthesizes CSD and returns it and that lead to some
failures.

Bug: 333806101
Test: atest CtsMediaDecoderTestCases:AdaptivePlaybackTest

Merged-In: If75c5b32060377e6342024d983d9c88d860b40e8
Change-Id: I8e68860971051dd1215f0740a5a9d758492d8b46