Fix StartActivityAsUserTests on Automotive
CLONE profiles are not supported on Automotive devices. Use a full
secondary user instead of a profile user to start the test activity on
Automotive. This will enable the tests for non-debug builds as well.

Also set the showForAllUsers flag to true for
StartActivityAsUserActivity, so that the activity can start for
non-current user.

The change should not affect phones. Tested for both cf_x86_64_phone
 and cf_x86_64_auto.

Fixes: 329662305

Test: atest CtsWindowManagerDeviceActivity:android.server.wm.activity.StartActivityAsUserTests

Change-Id: I95e46d344b237b2f46b42de4b17e1be8147ddf02