Fix ApplicationRestrictionsTest.
There was a bug where it would skip the private profile test because
it'd try to add the work profile on the private profile which is
obviously incorrect.

Additionally, the parameterised test was incorrect as it only tested the
work profile -> parent relationship rather than the work profile ->
private profile one.

Finally, the INTERACT_ACROSS_USERS issue mentioned on the bug isn't a
bug - the role actually does hold INTERACT_ACROSS_USERS.

Test: btest a.d.c.ApplicationRestrictionsTest
Bug: 335588148
Change-Id: Ief1378b2110d8f23b928bc0765cd386b77f6104f