skip configuration changes for closeToSquareDisplay
If the width and height of the device are close to a square, and the device is 180 degrees, excluding cutout, navigation bar, and status bar, the calculated width > height. This results in the orientation being calculated as landscape, consistent with the previous configuration reported, leading to the relaunch not being executed, thus causing a failure in the CTS test.

Bug: 335581849

Change-Id: Ic584e87c21bc4674ccbe2144eef70fc6c20a606e
Signed-off-by: tanxiaoyan <tanxiaoyan@xiaomi.com>