Enable --secondary_user_on_secondary_display for CtsDrmTestCases
--secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices
that have config_multiuserVisibleBackgroundUsers enabled. The main use
case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) interacting with driver's display.

Bug: 335718793
Test: atest --user-type secondary_user_on_secondary_display CtsDrmTestCases
Change-Id: I5a9517a2001e528dbd53f7148a525b3ae1668265