audio: Fix AudioFocusTest instant mode
The toggle notification access util asserts that the app in question
loses capabilities. If the app has shell permissions, and bedstead
doesn't appropriately contain them, then this assertation will fail.

Drop shell permissions before we do any of the state restoration, to
fulfill the preconditions of these utilities in a less brittle way.

Fixes: 332581642
Test: atest AudioFocusTest
Change-Id: Ic31cde6587716c70df333d5bf4f88d204ac10486