ITS:Cap the max preview test size at 1920x1440
Currently if the device does not support 4K resolution then we use
max preview size from the supported size.
This change caps the max preview testing size to 1920x1440 if the device
does not support the 4K resolution.

Test: Verified locally that if the device does not support 4K then the
max preview test size returned is 1920x1440 if the device supports it.

Bug: 329352025
Change-Id: I5722e6288c451d9db6164340a2ac56b9978b7070