Improvements to JobThrottlingTest
Initializing as many fields in the constructor to avoid hitting NPE in
the tearDown in case an assertion fails in setUp. This removes any
extraneous error messages that mask the original cause of test failure.

Failing early if WifiManager returns UNKNOWN_SSID.

Added the necessary permission for WifiManager to return the currently
connected SSID.

Removed some dead-code and moved to well established util methods for
polling checks.

Test: CtsJobSchedulerTestCases:JobThrottlingTest

Bug: 330263167
Bug: 330130241
Bug: 336667417
Bug: 335171431
Change-Id: I20b6ce225384c94f186f3eba7ba4153e3a4e5df3