Introduce ImeSettings.Builder#setTargetUser() for MockIme
This is a preparation for us to have some end-to-end tests for the IME
lifecycle event callbacks upon user switching (Bug 336658847). Such
end-to-end tests are critical for us to introduce per-user instance of
InputMethodBindingController (Bug 325515685) without breaking existing
behaviors.

In order to be able to easily write such tests, this CL makes it
possible for MockIme to be used to test cross-user scenarios as
follows.

  final Context contect = ...;
  final UiAutomation uiAutiomation = ...;
  final var settings1 = new ImeSettings.Builder();
  final var settings2 = new ImeSettings.Builder()
          .setTargetUser(UserHandle.of(secondaryUser));

  try (var session1 = MockImeSession.create(
          context, uiAutiomation, settings1);
       var session2 = MockImeSession.create(
          context, uiAutiomation, settings2)) {
    var stream1 = session1.openEventStream();
    var stream2 = session2.openEventStream();

    // interact with session1

    var extras = Map.of(
            MockTestActivityUtil.EXTRA_KEY_PRIVATE_IME_OPTIONS,
            marker);

    // switch to secondaryUser
    try (AutoCloseable ignored = MockTestActivityUtil.launchSyncAsUser(
            secondaryUser, isInstantApp(), extras, null)) {

        // interact with session2
        expectEvent(stream2, editorMatcher("onStartInput", marker),
                    TIMEOUT);
    }

    // describe tests here.
  }

Note that this no new test built on top of the above feature is
introduced yet. There must be no impact on existing single user tests.

The above functionality should also be helpful for us to convert
remaining host-side tests into device-side tests (Bug 323251870).

Bug: 336658847
Test: presubmit
Change-Id: Ib23b81ace9dc4e3b201b347e943f8b01f42a4193