Enable --secondary_user_on_secondary_display for CtsContentProviderTestsWithVisibility and CtsContentProviderTestsWithoutVisibility
--secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices
that have config_multiuserVisibleBackgroundUsers enabled. The main use
case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) interacting with driver's display.

Bug: 336674098
Bug: 336674200
Test: atest --user-type secondary_user_on_secondary_display CtsContentProviderTestsWithVisibility CtsContentProviderTestsWithoutVisibility
Change-Id: Ib9201741c47d51cf5efc4b0f865f1dcd9c1e798a