Fix CtsViewTestCases for secondary user on secondary display.
secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

1. Fix VerifyInputEventTest to check the display id assigned to
   a visible background user for input events.
2. Add secondary_user_on_secondary_display metadata.

Bug: 336963466
Test: atest --user-type secondary_user_on_secondary_display CtsViewTestCases
Test: atest CtsViewTestCases

Change-Id: I83593c4250d94e64629cea9d8bb196da3296f069
Signed-off-by: Hyunjae Choi <hyunjae.choi@lge.com>
Signed-off-by: Dohyun Lee <dohyun.lee@lge.com>