Allow activity starts for text CTS tests run in the sandbox
Starting activities in the SDK sandbox are normally not allowed.
However, for the purposes of CTS testing, they can be temporarily
enabled to allow the test to proceed. This requires a
START_ACTIVITIES_FROM_SDK_SANDBOX permission. Modify any CTS tests that
fail when run in the sandbox mode to allow the activity to be started.

Additionally, mark the activity as exported so that the sandbox process
can see this activity.

Any tests that are not relevant to the sandbox are annotated with
AppModeNonSdkSandbox.

Bug: 319428537
Test: atest <test-name> -- --enable-optional-parameterization --enable-parameterized-modules --module-parameter run_on_sdk_sandbox
Change-Id: I9f8d67b12caf5b1fbb247fe19849b743f170bbe8