Enable --secondary_user_on_secondary_display for CtsDpiTestCases and CtsDpiTestCases2
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 338082750
Test: Tested on gcar_md_x86_64 (AAOS Multi Display target)
Test: atest CtsDpiTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsDpiTestCases
Test: atest CtsDpiTestCases2
Test: atest --user-type secondary_user_on_secondary_display
CtsDpiTestCases2

Change-Id: I371536390f67ec0ffb9db81845062945f3cdbb9d