Added sleep to wait for the display to turn ON
This addresses the flake where oocasionally display would come up, but
the test would fail because no brightness event was emitted when
asserted. This is because WAKE_KEY event is routed via input, which
requests the display to wakeup, but doesnt necessarily wait for the
activity to complete.

Bug: 319086260
Test: atest
android.cts.statsdatom.display.DisplayWakeReportedStatsTests#testDisplayWakeReportedFromWakeKey
--iterations 500

Change-Id: If033bbb55eb740105a459b682689f2dc4f5d382e