CameraITS: Fix test_aspect_ratio_and_crop issue.
In this test, a check for supported stream combinations has been
introduced which prepares ImageReaders. While creating these
ImageReaders physicalCamera id is not passed in the command. After this
check in doCapture while preparing the ImageReaders the physicalCamera
id is also dropped creating the issues for tests with sub cameras.

Added physicalCamera id to the output surfaces in its_session_utils
before passing it to the command.

Bug: 336955126
Test: Test passes on pixel device for both camera id 0 and 0.3
Change-Id: I6df9f32e07851d722dc9038f5983d5f399afeb67