Improve connectivity job test correctness and latency
The scheduled job has no latency constraint so it may be started
immediately, and if it is killed after this, it will not be able to
report change in job state to TestAppInterface. Making sure to reset
this state at the time of the kill to avoid reading stale state.

Replacing TestUtils with PollingCheck as the former uses exponential
backoff while polling increasing the test duration significantly.

Enabling debug logs from network policy for the helper app to enable
simpler debugging for test failures in the future.

Test: atest CtsJobSchedulerTestCases:ConnectivityConstraintTest

Bug: 337953360
Change-Id: I60175ead5b994e6a2d4c2eaa91ef16c5269d9d39