Add a regression test for Bug 332912075
This end-to-end test is for the corresponding Framework change [1],
which is to fix an app compat issue introduced in my previous CL [2].

Suppose there are two EditText objects editText1 and editText2, and an
IME is shown for the editText1.

Whether the IME remains to be visible or not in the following code had
changed with my CL [2].

   editText2.requestFocus();
   editText1.setEnabled(false);

Before my CL [2] the IME remained to be shown, while it now becomes
hidden.

This CL works as an end-to-end test that the above scenario will not
be regressed again.

 [1]: I8dc003a38e7b95978fec83d055e44980163b70fb
 [2]: I01e5a29c0f82d068751774a2c58a1bd7b7b7f91a
      2a757ef4aea5cf9674be508e87378f66874ef163

Bug: 296466613
Bug: 332912075
Test: atest CtsInputMethodTestCases:EditTextImeSupportTest#testDisableImeAfterFocusChange
Change-Id: Ibc28a3b5c3ce987672f31a4125ab9204750dc530