Add voice assistant setup for non-default devices
Currently
WifiSliceTest#wifiSlice_grantedPermissionToDefaultVoiceAssistant
requires manual setup for devices with alternative voice
assistants (that are not com.google.android.googlequicksearchbox).
For these devices a manual voice assistant setup is required and
this CL removes the need for manual setup.

This CL is a copy of be808aa52655ba8d328383848a97641bd3fbc039
which does the same for wifiSlice_grantedPermissionToDefaultAssistant.

Bug: 339091165
Test: Test now passed without manual setup.
Change-Id: Id5f59e498d025fbb93a3ab3c68b27710bed56954