Ensure user control disabled clears stopped state
Also minor fixes to nearby tests - it is important that
setUserControlDisabled is removed before the package is uninstalled, but
it is not important that the package is stopped. So, wrapped the code
into try-finally and removed the misleadingly named method - its real
purpose was to remove the policy, not to stop the package.

Also updated the Package code to use isPackageStopped which is simpler
than checking FLAG_STOPPED.

Bug: 330688482
Test: btest CtsDevicePolicyTestCases:android.devicepolicy.cts.UserControlDisabledPackagesTest
Change-Id: I8613a90a7729ad2c0e23c7b978d220d939686c09