Fixes the calls to Thread.sleep(...) in tests
Thread.sleep(...) requires a non-negative argument. The tests fix a
timestamp in the future, then sleep for a difference of a time point
and that timestamp. This can be both positive or negative, depending
on how the test plays out. Which means, it may sometimes fail.

This change does not break existing passing tests.

Test: compiles
Bug: 335911344
Change-Id: I6e68cbafe411479a8626559c97357202c258f684