RESTRICT AUTOMERGE Prevent configuration change in WaitForResultActivity
We've observed (possible) configuration changes happen in
WaitForResultActivity, which does not handle them properly.

We cannot prevent configuration changes in every circumstance. But, this
change prevents *many* cases, and throws an exception in the other
cases. So, if a configuration change does happen after this, it will
be obvious, instead of some other obscure error.

Bug: 338042265
Test: presubmit
Change-Id: I33d873b698f0bc4c4a3eb1d815020a85efcf9db3