chore(onefingerpan): disable one finger pan in cts
In cts/FullScreenMagnificationGestureHandlerTest it would test if the touch
events are correctly propogated/delegated to view when magnification is
on. But if the one finger pan is enabled, the
FullScreenMagnificationGestureHandler will consume the events, then the
tests will fail. Therefore, we disable the one finger in
cts to keep the original tests healthy for now, and in the future we should add more test cases to ensure the tests cover different scenarios.

Bug: 330192193
Flag: ACONFIG com.android.server.accessibility.enable_magnification_one_finger_panning_gesture DEVELOPMENT
Test: atest cts/FullScreenMagnificationGestureHandlerTest
Change-Id: I5a56182201ae499f53d0eb76b2be203847fd4834