Higher-level classes for uinput devices
Before this CL, a user of UinputTouchDevice needed to also own a json
file in order to provide the commands that create the device of
interest, like touchscreen or stylus. That meant that it was necessary
to copy these json files to each build target that needed it. In most cases,
however, the user doesn't actually care
about the touchscreen properties; it's only important that the device is
created for a specific display, and behaves like a typical touchscreen.

To facilitate the usage of UinputTouchDevice, this CL introduces new
classes: UinputTouchScreen, UinputStylus, and the like.

These classes create the objects of the device type of interest, without
requiring the user to be aware of the details of the json files.
Eventually, these can be expanded to provide device-specific
functionality, like the ability to send higher-level events.

Additionally, prior to this CL, it was standard to modify the json
configuration in order to override the device properties. We are now
introducing a "RegisterCommand", a data class that will contain these
values instead. This allows simpler manipulation of the device
properties, since we would now be interacting with kt code rather than
manipulating json values at runtime. We can't fully deprecate the json
files yet, since the "uinput" command relies on them. However, maybe
this can be achieved one day by further refactoring or removing the
uinput command.

Bug: 329439551
Test: atest CtsInputTestCases
Change-Id: I86d5b99e4f97ceff1de4c29e845975ed148a4895