Use uinput-based injection to make tests hermetic
Make WallpaperWindowInputTests hermetic by switching to
uinput-based injection. That means that the system will be left in
a clean state when the test completes execution, even if it fails.

At the same time, update other tests to use the new infra developed
here. The eventual goal is to only expose the higher-level APIs to the
users.

Bug: 334214043
Test: atest CtsWindowManagerDeviceInput
Change-Id: I74487fc7796a5859d6e9a8bba207e5763e409d6c