Camera: Add runtime flag checks before accessing flagged symbols.
CameraDeviceInfo.mako currently unconditionally references flagged
keys which may lead to situations where it refers to keys that
are stripped from the system it is built for.

I74ed48eb33 updated CameraDeviceInfo.mako to add runtime flag
checking. This patch is the change from running
metadata-generate on with the new logic.

Bug: 338003995
Test: Manually verified that the generated source
      correctly checks the flags.
Change-Id: I5a369647e4d97adf986cc32774d40f2aad1479d2