Temporarily disable cts tests failing due to desktop mode
testMetricsMatchesLayoutOnFreeformActivity and
testPreQTopProcessResumedActivityInFreeform are failing specifically
when desktop mode is enabled on the device. We have a fix pending
(ag/27360596 and ag/27344155) but it will not land in time for the SLO
demoting the tests.

To prevent all the wm CTS tests from being demoted we will disable these
tests on devices that are running desktop mode and re-enable them once
the fixes land.

Test: atest CtsWindowManagerDeviceActivity:ActivityLifecycleFreeformTests
Test: atest CtsWindowManagerDeviceWindow:WindowMetricsActivityTests
Bug: 339969180
Bug: 339619498
Change-Id: I1d01d4f20f6908ae3eff0bcc92f9a41e4c201d6c