Wait for service enable/disable before continuing.
NonUiInCallServiceTest enables and disables the mock ICS related to the
tests a few times.  It does so by telling packagemanager to do this and
moving on.  When reviewing how PackageManager CTS tests are authored,
those tests use a broadcast receiver to get confirmation that the enable
or disable is completed before continuing on.  A similar approach is used
here so that we can be sure the enable/disable state of the ICS has
changed as required.  This is important because disabling a service will
kill the host package as well.  We could just flag to NOT kill the app,
however when we change the service state to disabled at the start of the
test the side effect of the app being killed is actually desirable to
ensure we're in a good state anyways.

Test: Re-run these tests to verify passing state.
Fixes: 321205341
Change-Id: I83455775a0f48615189ea4dcfaf11c4c7e74c4a7