mediav2 CTS: Reduce runtime of Decoder Tests
There were multiple test parameters which were required only
for reconfigure test but it was added to CodecDecoderTest
and CodecDecoderSurfaceTest parameters, so it was running for
other tests inside CodecDecoderTest and CodecDecoderSurfaceTest.
Move out these reconfigure parameters to a separate reconfigure
tests.
This change reduces runtime from ~16 min to ~5 min for CodecDecoderTest,
from ~13 min to ~5 min for CodecDecoderSurfaceTest.
CodecDecoderReconfigureTest takes ~4 min and
CodecDecoderSurfaceReconfigureTest takes ~3 min.

Bug: 342531342
Test: atest android.mediav2.cts.CodecDecoderTest
Test: atest android.mediav2.cts.CodecDecoderSurfaceTest
Test: atest android.mediav2.cts.CodecDecoderReconfigureTest
Test: atest android.mediav2.cts.CodecDecoderSurfaceReconfigureTest

Change-Id: I4a880ab707351c114aeb8403638526bdc868cfa1