Fix CtsWidgetTestCases.ScrollViewTest for watches
In ag/27170634, the top padding of the test screen was increased by a
value too high for watches, which meant that the tested ScrollView
sometimes was outside the boundaries of the screen.

This change provides a simple fix by reverting the padding value back to
the previous value for watches only, without affecting other
form-factors.

Bug: 341847822
Bug: 338320682
Test: atest ScrollViewTest
Change-Id: I456c78357cb07f78233958dad0ba39254c80e9eb