Remove global injection from CtsTouchUtils
...and use uinput for global injection in places that actually require
global injection, like SurfaceControlViewHostTests.

Currently, uiautomation is used to inject input events globally. That
provides room for error, because the caller may forget to clean up a
dangling pointer. That would typically cause subsequent tests to fail.

To avoid this issue, we convert the tests that need global injection to
use uinput-based approach. At the end of the test, the uinput device is
cleaned up, which will make the tests hermetic.

The eventual goal is to avoid calling 'injectInputEvent' completely.

Bug: 329439551
Test: atest SurfaceControlViewHostTests
Test: atest CtsSurfaceControlTests
Test: m CtsSurfaceControlTests CtsViewTestCases CtsWidgetTestCases
Change-Id: Idf2d53bc1f4ff04ea8381af913edbbdf186c19bd