Resolve ExtendedInCallServiceTest flakiness
Update the internal call audio state before updating the system audio
route in CallAudioRouteStateMachine. In the tests, we're seeing that
when requesting an endpoint change, it gets ignored because the
EndpointController is retrieving the current call audio route to
determine whether or not the update would be a stale request. Because
the existing code updates the internal call audio state after the system
audio update, it's possible that by the time the request to change the
endpoint is received, the internal call audio state hasn't been updated
even if the audio route has transitioned.

Also, ensure that we wait for the connection's endpoint to be populated
by waiting for the onCallEndpointChanged callback to be invoked instead
of only waiting for the ICS's callback in the existing test. We update
the connection service's endpoint after so the verification can happen
before the connection service receives its update.

Bug: 340793169
Test: atest ExtendedInCallServiceTest
Change-Id: I277952883a62b604cd3814592fab4fb21c1216b1