Enable --secondary_user_on_secondary_display for CtsMediaDrmFrameworkTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices
that have config_multiuserVisibleBackgroundUsers enabled. The main use
case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) interacting with driver's display.

Bug: 335718630
Test: atest --user-type secondary_user_on_secondary_display CtsMediaDrmFrameworkTestCases
Test: atest --user-type secondary_user CtsMediaDrmFrameworkTestCases
Test: atest CtsMediaDrmFrameworkTestCases
Change-Id: Iaf23a133ded25d903ede505f3d88f2f44dbb9fb5