mediav2: Ensure VP9 encoder doesn't generate CSD
If VP9 encoder generates CSD and if that data is sent to the
MediaMuxer via writeSampleData() API, then then the generated
file is invalid and can't be played back.

VP9 encoders so far didn't generate CSD and this check ensures
that the current behavior continues till MediaMuxer is updated to
handle such input to writeSampleData().

Bug: 339166419
Bug: 332052663

Test: atest CtsMediaV2TestCases:CodecEncoderTest#testSimpleEncode

Change-Id: Ie28ee2670d35afde8f2533abaa3e359501c552e8