Enable secondary_user_on_secondary_display for CtsMediaCodecTestCases
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 342505726

Test: atest --user-type secondary_user_on_secondary_display \
          CtsMediaCodecTestCases
Test: atest CtsMediaCodecTestCases

Change-Id: I2709f248aa6f24b642f648dbc3e82548664bc9ab